Rename namespace MahApps.Metro.Behaviours to MahApps.Metro.Behaviors
