+++ src/MahApps.Metro.Samples/MahApps.Metro.Demo/ExampleViews/ColorExample.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:MetroDemo="clr-namespace:MetroDemo"
-             xmlns:behaviours="http://metro.mahapps.com/winfx/xaml/shared"
+             xmlns:behaviors="http://metro.mahapps.com/winfx/xaml/shared"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
              xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
@@ -31,7 +31,7 @@
                                  Height="30"
                                  Stroke="{DynamicResource GrayBrush2}">
                             <i:Interaction.Behaviors>
-                                <behaviours:BindableResourceBehavior ResourceName="{Binding}" Property="{x:Static Shape.FillProperty}" />
+                                <behaviors:BindableResourceBehavior ResourceName="{Binding}" Property="{x:Static Shape.FillProperty}" />
                             </i:Interaction.Behaviors>
                         </Ellipse>
                         <TextBlock Grid.Column="1"
