+++ src/MahApps.Metro/Behaviors/ReloadBehavior.cs
@@ -3,7 +3,7 @@ using System.Windows;
 using System.Windows.Controls;
 using MahApps.Metro.Controls;
 
-namespace MahApps.Metro.Behaviours
+namespace MahApps.Metro.Behaviors
 {
     public static class ReloadBehavior
     {
