+++ src/MahApps.Metro/Controls/Dialogs/LoginDialog.xaml
@@ -1,7 +1,7 @@
 ﻿<Dialogs:BaseMetroDialog x:Class="MahApps.Metro.Controls.Dialogs.LoginDialog"
                          xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                          xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                         xmlns:Behaviors="clr-namespace:MahApps.Metro.Behaviours"
+                         xmlns:Behaviors="clr-namespace:MahApps.Metro.Behaviors"
                          xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
                          xmlns:Dialogs="clr-namespace:MahApps.Metro.Controls.Dialogs">
     <Grid Margin="0 10 0 0">
