+++ src/MahApps.Metro/Properties/AssemblyInfo.cs
@@ -15,7 +15,7 @@ using System.Windows.Markup;
 [assembly: XmlnsPrefix(@"http://metro.mahapps.com/winfx/xaml/controls", "mah")]
 [assembly: XmlnsPrefix(@"http://metro.mahapps.com/winfx/xaml/shared", "mah")]
 
-[assembly: XmlnsDefinition(@"http://metro.mahapps.com/winfx/xaml/shared", "MahApps.Metro.Behaviours")]
+[assembly: XmlnsDefinition(@"http://metro.mahapps.com/winfx/xaml/shared", "MahApps.Metro.Behaviors")]
 [assembly: XmlnsDefinition(@"http://metro.mahapps.com/winfx/xaml/shared", "MahApps.Metro.Actions")]
 [assembly: XmlnsDefinition(@"http://metro.mahapps.com/winfx/xaml/shared", "MahApps.Metro.Converters")]
 [assembly: XmlnsDefinition(@"http://metro.mahapps.com/winfx/xaml/controls", "MahApps.Metro")]
