+++ src/MahApps.Metro/Styles/Controls.AnimatedSingleRowTabControl.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:Behaviours="clr-namespace:MahApps.Metro.Behaviours"
+                    xmlns:Behaviors="clr-namespace:MahApps.Metro.Behaviors"
                     xmlns:Controls="clr-namespace:MahApps.Metro.Controls">
 
     <ResourceDictionary.MergedDictionaries>
@@ -178,7 +178,7 @@
             </ScrollViewer>
             <Controls:TransitioningContentControl x:Name="ContentPanel"
                                                   Grid.Row="1"
-                                                  Behaviours:ReloadBehavior.OnSelectedTabChanged="True"
+                                                  Behaviors:ReloadBehavior.OnSelectedTabChanged="True"
                                                   RestartTransitionOnContentChange="True"
                                                   Transition="{TemplateBinding Controls:TabControlHelper.Transition}"
                                                   UseLayoutRounding="True">
@@ -239,7 +239,7 @@
             </ScrollViewer>
             <Controls:TransitioningContentControl x:Name="ContentPanel"
                                                   Grid.Column="0"
-                                                  Behaviours:ReloadBehavior.OnSelectedTabChanged="True"
+                                                  Behaviors:ReloadBehavior.OnSelectedTabChanged="True"
                                                   RestartTransitionOnContentChange="True"
                                                   Transition="{TemplateBinding Controls:TabControlHelper.Transition}"
                                                   UseLayoutRounding="True">
