+++ src/MahApps.Metro/Styles/Controls.DatePicker.xaml
@@ -2,7 +2,7 @@
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                     xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
                     xmlns:Converters="clr-namespace:MahApps.Metro.Converters"
-                    xmlns:behaviours="clr-namespace:MahApps.Metro.Behaviours"
+                    xmlns:behaviors="clr-namespace:MahApps.Metro.Behaviors"
                     xmlns:i="http://schemas.microsoft.com/xaml/behaviors">
 
     <ResourceDictionary.MergedDictionaries>
@@ -75,7 +75,7 @@
                                                Foreground="{TemplateBinding Foreground}"
                                                SelectionBrush="{DynamicResource HighlightBrush}">
                                 <i:Interaction.Behaviors>
-                                    <behaviours:DatePickerTextBoxBehavior />
+                                    <behaviors:DatePickerTextBoxBehavior />
                                 </i:Interaction.Behaviors>
                             </DatePickerTextBox>
