+++ src/MahApps.Metro/Themes/DateTimePicker.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:Behaviours="clr-namespace:MahApps.Metro.Behaviours"
+                    xmlns:Behaviors="clr-namespace:MahApps.Metro.Behaviors"
                     xmlns:Controls="clr-namespace:MahApps.Metro.Controls"
                     xmlns:Converters="clr-namespace:MahApps.Metro.Converters"
                     xmlns:i="http://schemas.microsoft.com/xaml/behaviors"
@@ -160,7 +160,7 @@
                                                IsReadOnly="{Binding Path=IsReadOnly, RelativeSource={RelativeSource TemplatedParent}}"
                                                SelectionBrush="{DynamicResource HighlightBrush}">
                                 <i:Interaction.Behaviors>
-                                    <Behaviours:DatePickerTextBoxBehavior />
+                                    <Behaviors:DatePickerTextBoxBehavior />
                                 </i:Interaction.Behaviors>
                             </DatePickerTextBox>
                             <ContentControl x:Name="PART_FloatingMessageContainer"
